requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.0
jinja2==3.1.2
SQLAlchemy==1.4.46
email-validator==2.0.0
Pillow
pytest==7.3.1
pyautogui==0.9.53
psycopg2
gunicorn==20.1.0
requests==2.28.1
beautifulsoup4==4.11.1

